+++ src/main/kotlin/com/github/alaanor/candid/util/PathUtil.kt
@@ -4,17 +4,23 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import java.nio.file.Paths
 
-fun PsiElement.filePath(): String = this.containingFile.originalFile.virtualFile.path
-fun PsiElement.filePathDirectory(): String = this.containingFile.originalFile.virtualFile.parent.path
+fun PsiElement.filePath(): String = containingFile.originalFile.virtualFile.path
+fun PsiElement.filePathDirectory(): String = containingFile.originalFile.virtualFile.parent.path
+
+fun PsiElement.projectFilePath(): String {
+    return Paths.get(project.basePath)
+        .relativize(Paths.get(containingFile.originalFile.virtualFile.path))
+        .toString()
+}
 
 fun PsiElement.getRelativePath(psiElement: PsiElement): String {
-    return Paths.get(this.filePathDirectory())
+    return Paths.get(filePathDirectory())
         .relativize(Paths.get(psiElement.filePath()))
         .toString()
 }
 
 fun PsiElement.getRelativePath(virtualFile: VirtualFile): String {
-    return Paths.get(this.filePathDirectory())
+    return Paths.get(filePathDirectory())
         .relativize(Paths.get(virtualFile.path))
         .toString()
 }