+++ src/main/kotlin/com/github/alaanor/candid/inspection/CandidSelfImportInspection.kt
@@ -1,6 +1,7 @@
 package com.github.alaanor.candid.inspection
 
 import com.github.alaanor.candid.psi.CandidImportStatement
+import com.github.alaanor.candid.psi.getTextRangeWithoutQuote
 import com.github.alaanor.candid.psi.importedPsiFile
 import com.github.alaanor.candid.quickfix.CandidRemoveImportQuickFix
 import com.github.alaanor.candid.util.projectFilePath
@@ -8,7 +9,7 @@ import com.intellij.codeInspection.*
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiTreeUtil
 
-class CandidSelfImportInspection: LocalInspectionTool() {
+class CandidSelfImportInspection : LocalInspectionTool() {
 
     override fun checkFile(file: PsiFile, manager: InspectionManager, isOnTheFly: Boolean): Array<ProblemDescriptor>? {
         val problemsHolder = ProblemsHolder(manager, file, isOnTheFly)
@@ -22,7 +23,7 @@ class CandidSelfImportInspection: LocalInspectionTool() {
                     import,
                     "Cannot self import",
                     ProblemHighlightType.GENERIC_ERROR,
-                    import.stringLiteral?.textRange ?: import.textRange,
+                    import.stringLiteral?.getTextRangeWithoutQuote() ?: import.textRange,
                     CandidRemoveImportQuickFix(import, CandidRemoveImportQuickFix.Type.Invalid)
                 )
             }