+++ src/main/kotlin/com/github/alaanor/candid/psi/mixin/CandidIdentifierDeclarationMixin.kt
@@ -1,5 +1,6 @@
 package com.github.alaanor.candid.psi.mixin
 
+import com.github.alaanor.candid.icon.CandidIcons
 import com.github.alaanor.candid.psi.CandidElementFactory
 import com.github.alaanor.candid.psi.CandidIdentifierDeclaration
 import com.github.alaanor.candid.psi.primitive.CandidNameIdentifierOwner
@@ -7,17 +8,20 @@ import com.github.alaanor.candid.psi.primitive.CandidStubBasedElementBase
 import com.github.alaanor.candid.psi.stub.CandidStubBasedPsiElement
 import com.github.alaanor.candid.psi.stub.impl.CandidIdentifierDeclarationStub
 import com.intellij.lang.ASTNode
+import com.intellij.navigation.ItemPresentation
 import com.intellij.psi.PsiElement
 import com.intellij.psi.stubs.IStubElementType
+import javax.swing.Icon
 
 abstract class CandidIdentifierDeclarationMixin :
     CandidStubBasedElementBase<CandidIdentifierDeclarationStub>,
     CandidNameIdentifierOwner,
     CandidStubBasedPsiElement<CandidIdentifierDeclarationStub, CandidIdentifierDeclaration>,
-    CandidIdentifierDeclaration {
+    CandidIdentifierDeclaration,
+    ItemPresentation {
 
     constructor(node: ASTNode) : super(node)
-    constructor(stub: CandidIdentifierDeclarationStub, type: IStubElementType<*, *>): super(stub, type)
+    constructor(stub: CandidIdentifierDeclarationStub, type: IStubElementType<*, *>) : super(stub, type)
 
     override fun getNameIdentifier(): PsiElement = node.psi
     override fun getTextOffset(): Int = node.startOffset
@@ -27,4 +31,7 @@ abstract class CandidIdentifierDeclarationMixin :
         nameIdentifier.replace(CandidElementFactory.createTypeDeclaration(project, name))
         return this
     }
+
+    override fun getIcon(unused: Boolean): Icon = CandidIcons.Type
+    override fun getPresentableText(): String = text
 }